fix(classification/f12): build/test hardening round — span brace literals, qsizetype casts, OpenCV 4 API, e2e fixture realism
- test fixtures: named vectors for std::span parameters (spans have no
  initializer_list constructor), exact-binary fractions (3/4, 1/4) for
  hand-computed ECE bin math, N×K score-matrix layouts, realistic NB
  training jitter (zero-variance training overflows PDFs to NaN)
- RTrees feature importance: getVarImportance (OpenCV 4 API)
- RF backend for uncertainty E2E: OpenCV NB predictProb underflows to
  all-zero rows for off-distribution pixels (pre-existing ADR 0094
  fragility — now documented; the -1 uncertainty sentinel exposes it
  honestly instead of reporting fake confidence)
- object postprocess tests: pre-sized output rasters (typed SizeMismatch)
- all 17 targeted classification suites green (local evidence, pass 1)

+++ src/app/workbench/classification_studio_widget.cpp
@@ -439,7 +439,7 @@ void RsProbabilityPanel::paintEvent( QPaintEvent *event )
   painter.fillRect( rect(), QColor( 12, 12, 16 ) );
   if ( mClassIds.isEmpty() )
     return;
-  const int rowH = std::max( 12, height() / std::max( 1, mClassIds.size() ) );
+  const int rowH = std::max( 12, height() / std::max( 1, static_cast<int>( mClassIds.size() ) ) );
   int y = 0;
   for ( int i = 0; i < mClassIds.size(); ++i )
   {
@@ -574,7 +574,7 @@ void RsConfusionPairsWidget::paintEvent( QPaintEvent *event )
   int maxCount = 1;
   for ( int c : mCounts )
     maxCount = std::max( maxCount, c );
-  const int rowH = std::max( 12, height() / std::max( 1, mCounts.size() ) );
+  const int rowH = std::max( 12, height() / std::max( 1, static_cast<int>( mCounts.size() ) ) );
   int y = 0;
   for ( int i = 0; i < mCounts.size(); ++i )
   {
@@ -637,7 +637,7 @@ void RsFeatureImportanceWidget::paintEvent( QPaintEvent *event )
     maxV = std::max( maxV, v );
   if ( maxV <= 0.0 )
     maxV = 1.0;
-  const int rowH = std::max( 12, height() / std::max( 1, mImportances.size() ) );
+  const int rowH = std::max( 12, height() / std::max( 1, static_cast<int>( mImportances.size() ) ) );
   int y = 0;
   for ( int i = 0; i < mImportances.size(); ++i )
   {
